requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bleach==5.0.1
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
dataclasses==0.6
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.28
gotrue==0.5.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==2.10
importlib-metadata==5.0.0
iniconfig==1.1.1
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.9.3
MarkupSafe==2.1.1
more-itertools==8.14.0
packaging==21.3
pkginfo==1.8.3
pluggy==1.0.0
postgrest-py==0.10.2
protobuf==4.21.7
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.10.0
python-semantic-release==7.28.1
readme-renderer==37.2
realtime==0.0.4
realtime-py==0.1.3
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.3
storage3==0.3.4
supabase==0.5.8
supabase-py==0.0.2
toml==0.10.2
tomlkit==0.10.2
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.2
zipp==3.8.1
